master/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
APScheduler==3.6.3
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
config==0.4.2
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Executor==0.9.3
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==1.0.0
zipp==3.1.0
